Rework wildcard workflow into 4 phases
Refactor the Wildcard recon workflow into an explicit 4-phase structure (asset_discovery, validation, content_discovery, vulnerability_scanning). Move historical URL discovery (Wayback/Gau) from asset discovery into Phase 3 and implement a new content_discovery.go with steps for URL discovery, crawling, JS analysis, subdomain mini re-probe, param discovery, consolidation, and ffuf. Rename/passive_enum.go -> asset_discovery.go, update step numbering and logger labels, and ensure WildcardSteps ordering/comments in pkg/scan/scan.go match the new phases. Update CLI help and skills docs to reflect the phase model and resume/state behavior notes. Simplify and consolidate export outputs in pkg/utils/export.go (change file formats and reduce duplicate files), and apply minor imports/order fixes in setup code.
