+++ pkg/setup/gf_patterns.go
@@ -8,8 +8,8 @@ import (
 	"encoding/json"
 	"fmt"
 	"os"
-	"path/filepath"
 	"os/exec"
+	"path/filepath"
 
 	"github.com/vishnu303/chaathan-flow/pkg/progress"
 )
