+++ pkg/wildcard_flow/validation.go
@@ -1,0 +1,207 @@
+// Phase 2 — Validation & Fingerprint (Steps 5–9)
+//
+// Validates discovered assets through DNS resolution, HTTP probing,
+// TLS analysis, and port scanning.
+//
+//  5. Consolidation & DNS Resolution (DNSx)
+//  6. DNS Brute-force (ShuffleDNS) [Optional]
+//  7. Live Web Server Probing (Httpx)
+//  8. TLS Certificate Analysis (tlsx) + host metadata enrichment [Optional]
+//  9. Port Scanning (Naabu) [Optional]
+package wildcard_flow
+
+import (
+	"context"
+
+	"github.com/vishnu303/chaathan-flow/pkg/logger"
+	"github.com/vishnu303/chaathan-flow/pkg/metadata"
+	"github.com/vishnu303/chaathan-flow/pkg/utils"
+)
+
+// ─────────────────────────────────────────────────────────────
+// Step 5 — Consolidation & DNS Resolution
+// ─────────────────────────────────────────────────────────────
+
+// stepDNSConsolidation merges all passive sources and resolves them with DNSx.
+// Returns true if the scan should be cancelled.
+func stepDNSConsolidation(c *Ctx) bool {
+	if c.State.IsStepCompleted("dns_resolution") {
+		logger.Section("Step 6: Consolidation & DNS Resolution [RESUMED — skipping]")
+		return c.cancelled()
+	}
+	logger.Section("Step 6: Consolidating Subdomains")
+
+	passiveSources := existingFiles(
+		c.F.SubfinderOut,
+		c.F.AssetfinderOut,
+		c.F.Sublist3rOut,
+		c.F.AmassOut,
+		c.F.GithubSubsOut,
+		c.F.SubdomainizerOut,
+	)
+	if err := utils.MergeAndDeduplicate(passiveSources, c.F.ConsolidatedSubs); err != nil {
+		c.StateMgr.MarkStepFailed(c.State, "dns_resolution", err)
+		logger.Error("Failed to consolidate: %v", err)
+	}
+	logger.Success("Consolidated list saved to %s", c.F.ConsolidatedSubs)
+
+	logger.SubStep("Running DNSx for resolution...")
+	if err := c.Tb.RunDnsx(c.GoCtx, c.F.ConsolidatedSubs, c.F.DnsxOut); err != nil {
+		c.StateMgr.MarkStepFailed(c.State, "dns_resolution", err)
+		logger.Error("DNSx failed: %v", err)
+	}
+	c.StateMgr.MarkStepComplete(c.State, "dns_resolution")
+	return c.cancelled()
+}
+
+// ─────────────────────────────────────────────────────────────
+// Step 6 — DNS Brute-force (ShuffleDNS)
+// ─────────────────────────────────────────────────────────────
+
+// stepDNSBruteforce runs ShuffleDNS when a dns-wordlist is provided.
+// Returns true if the scan should be cancelled.
+func stepDNSBruteforce(c *Ctx) bool {
+	if c.State.IsStepCompleted("dns_bruteforce") {
+		logger.Section("Step 7: DNS Brute-force (ShuffleDNS) [RESUMED — skipping]")
+	} else if !c.SkipShuffleDNS && c.DNSWordlistPath != "" {
+		logger.Section("Step 7: DNS Brute-force (ShuffleDNS)")
+		logger.SubStep("Running ShuffleDNS with wordlist: %s", c.DNSWordlistPath)
+
+		if err := runWithSkip(c, "shuffledns", func(sCtx context.Context) error {
+			return c.Tb.RunShuffleDNS(sCtx, c.Domain, c.DNSWordlistPath, c.ResolversPath, c.F.ShufflednsOut)
+		}); err != nil {
+			if err == ErrToolSkipped {
+				logger.Info("  ShuffleDNS skipped")
+			} else {
+				c.StateMgr.MarkStepFailed(c.State, "dns_bruteforce", err)
+				logger.Warning("ShuffleDNS failed: %v", err)
+			}
+		} else {
+			if c.ScanID > 0 {
+				count, _ := utils.ParseSubdomainsFile(c.ScanID, c.F.ShufflednsOut, "shuffledns")
+				logger.Info("  Found %d subdomains via DNS brute-force", count)
+			}
+			// Merge brute-forced subs back into the consolidated list
+			utils.MergeAndDeduplicate(
+				[]string{c.F.ConsolidatedSubs, c.F.ShufflednsOut},
+				c.F.ConsolidatedSubs,
+			)
+		}
+		c.StateMgr.MarkStepComplete(c.State, "dns_bruteforce")
+	} else if c.SkipShuffleDNS {
+		logger.Section("Step 7: Skipping ShuffleDNS (--skip-shuffledns)")
+		c.StateMgr.MarkStepComplete(c.State, "dns_bruteforce")
+	} else {
+		logger.Section("Step 7: Skipping ShuffleDNS (no --dns-wordlist provided)")
+		logger.Info("Use --dns-wordlist to enable DNS brute-force")
+		c.StateMgr.MarkStepComplete(c.State, "dns_bruteforce")
+	}
+	return c.cancelled()
+}
+
+// ─────────────────────────────────────────────────────────────
+// Step 7 — Live Web Server Probing (Httpx)
+// ─────────────────────────────────────────────────────────────
+
+// stepHTTPProbing probes all consolidated subdomains with Httpx.
+// Returns true if the scan should be cancelled.
+func stepHTTPProbing(c *Ctx) bool {
+	if c.State.IsStepCompleted("http_probing") {
+		logger.Section("Step 8: Live Web Server Probing [RESUMED — skipping]")
+		return c.cancelled()
+	}
+	logger.Section("Step 8: Live Web Server Probing")
+	logger.SubStep("Running Httpx...")
+
+	if err := c.Tb.RunHttpx(c.GoCtx, c.F.ConsolidatedSubs, c.F.HttpxOut); err != nil {
+		c.StateMgr.MarkStepFailed(c.State, "http_probing", err)
+		logger.Error("Httpx failed: %v", err)
+	} else {
+		if c.ScanID > 0 {
+			count, _ := utils.ParseHttpxOutput(c.ScanID, c.F.HttpxOut)
+			logger.Info("  Found %d live hosts", count)
+		}
+	}
+	c.StateMgr.MarkStepComplete(c.State, "http_probing")
+	return c.cancelled()
+}
+
+// ─────────────────────────────────────────────────────────────
+// Step 8 — TLS Certificate Analysis (tlsx) + host metadata
+// ─────────────────────────────────────────────────────────────
+
+// stepTLSAnalysis examines TLS certificates and enriches host metadata.
+// Returns true if the scan should be cancelled.
+func stepTLSAnalysis(c *Ctx) bool {
+	if c.State.IsStepCompleted("tls_analysis") {
+		logger.Section("Step 9: TLS Certificate Analysis (tlsx) [RESUMED — skipping]")
+	} else if !c.SkipTlsx {
+		logger.Section("Step 9: TLS Certificate Analysis (tlsx)")
+		logger.SubStep("Running tlsx — extracting SANs and checking cert issues...")
+
+		if err := c.Tb.RunTlsx(c.GoCtx, c.F.ConsolidatedSubs, c.F.TlsxOut); err != nil {
+			c.StateMgr.MarkStepFailed(c.State, "tls_analysis", err)
+			logger.Warning("tlsx failed: %v", err)
+		} else {
+			if c.ScanID > 0 {
+				newSubs, certVulns, _ := utils.ParseTlsxOutput(c.ScanID, c.F.TlsxOut, c.Domain)
+				if newSubs > 0 {
+					logger.Info("  Discovered %d new subdomains from certificate SANs", newSubs)
+				}
+				if certVulns > 0 {
+					logger.Info("  Found %d certificate issues (expired/self-signed/mismatch)", certVulns)
+				}
+			}
+		}
+		c.StateMgr.MarkStepComplete(c.State, "tls_analysis")
+	} else {
+		logger.Section("Step 9: Skipping tlsx (--skip-tlsx)")
+		c.StateMgr.MarkStepComplete(c.State, "tls_analysis")
+	}
+
+	// Host metadata enrichment (always attempted after step 9)
+	if c.ScanID > 0 && utils.FileExists(c.F.HttpxOut) {
+		hostTargetCount := collectLiveHostTargetsFromHttpx(c.F.HttpxOut, c.F.HttpxLiveHosts)
+		if hostTargetCount > 0 {
+			logger.SubStep("Collecting lightweight host metadata for ROI scoring...")
+			hostTargets := loadLineSlice(c.F.HttpxLiveHosts, 250)
+			if count, err := metadata.CollectHostMetadata(c.ScanID, hostTargets); err != nil {
+				logger.Warning("Host metadata enrichment failed: %v", err)
+			} else if count > 0 {
+				logger.Info("  Stored metadata for %d live hosts", count)
+			}
+		}
+	}
+
+	return c.cancelled()
+}
+
+// ─────────────────────────────────────────────────────────────
+// Step 9 — Port Scanning (Naabu)
+// ─────────────────────────────────────────────────────────────
+
+// stepPortScanning runs Naabu against all discovered subdomains.
+// Returns true if the scan should be cancelled.
+func stepPortScanning(c *Ctx) bool {
+	if c.State.IsStepCompleted("port_scanning") {
+		logger.Section("Step 10: Port Scanning [RESUMED — skipping]")
+	} else if !c.SkipNaabu {
+		logger.Section("Step 10: Port Scanning")
+		logger.SubStep("Running Naabu on all discovered subdomains...")
+
+		if err := c.Tb.RunNaabuList(c.GoCtx, c.F.ConsolidatedSubs, c.F.NaabuOut); err != nil {
+			c.StateMgr.MarkStepFailed(c.State, "port_scanning", err)
+			logger.Error("Naabu failed: %v", err)
+		} else {
+			if c.ScanID > 0 {
+				count, _ := utils.ParseNaabuOutput(c.ScanID, c.F.NaabuOut)
+				logger.Info("  Found %d open ports", count)
+			}
+		}
+		c.StateMgr.MarkStepComplete(c.State, "port_scanning")
+	} else {
+		logger.Section("Step 10: Skipping Naabu (--skip-naabu)")
+		c.StateMgr.MarkStepComplete(c.State, "port_scanning")
+	}
+	return c.cancelled()
+}
