+++ pkg/wildcard_flow/vulnerability_scanning.go
@@ -1,7 +1,11 @@
-// Vulnerability Scanning — Steps 17–18
+// Phase 4 — Vulnerability Scanning (Steps 17–20)
+//
+// Runs security scanners against live hosts and URLs.
 //
 //  17. Vulnerability Scanning — Infrastructure (Nuclei on live httpx hosts)
 //  18. Vulnerability Scanning — URLs (Nuclei on filtered/GF-matched URLs)
+//  19. Subdomain Takeover Detection (Subjack) [Optional]
+//  20. XSS Scanning (Dalfox) on parameterised live URLs [Optional]
 package wildcard_flow
 
 import (
@@ -129,6 +133,95 @@ func stepVulnScanningURLs(c *Ctx) bool {
 }
 
 // ─────────────────────────────────────────────────────────────
+// Step 19 — Subdomain Takeover Detection (Subjack)
+// ─────────────────────────────────────────────────────────────
+
+// stepTakeoverDetection runs Subjack to find dangling CNAMEs.
+// Returns true if the scan should be cancelled.
+func stepTakeoverDetection(c *Ctx) bool {
+	if c.State.IsStepCompleted("takeover_detection") {
+		logger.Section("Step 19: Subdomain Takeover Detection (Subjack) [RESUMED — skipping]")
+	} else if !c.SkipSubjack {
+		logger.Section("Step 19: Subdomain Takeover Detection (Subjack)")
+		logger.SubStep("Running Subjack — checking for dangling CNAMEs...")
+
+		if err := c.Tb.RunSubjack(c.GoCtx, c.F.ConsolidatedSubs, c.F.SubjackOut); err != nil {
+			c.StateMgr.MarkStepFailed(c.State, "takeover_detection", err)
+			logger.Warning("Subjack failed: %v", err)
+		} else {
+			if c.ScanID > 0 {
+				count, _ := utils.ParseSubjackOutput(c.ScanID, c.F.SubjackOut)
+				if count > 0 {
+					logger.Success("  🚨 Found %d potential subdomain takeovers!", count)
+					// Notify immediately — takeovers are critical
+					if c.Notifier != nil {
+						sendTakeoverNotifications(c)
+					}
+				} else {
+					logger.Info("  No subdomain takeovers detected")
+				}
+			}
+		}
+		c.StateMgr.MarkStepComplete(c.State, "takeover_detection")
+	} else {
+		logger.Section("Step 19: Skipping Subjack (--skip-subjack)")
+		c.StateMgr.MarkStepComplete(c.State, "takeover_detection")
+	}
+	return c.cancelled()
+}
+
+// ─────────────────────────────────────────────────────────────
+// Step 20 — XSS Scanning (Dalfox)
+// ─────────────────────────────────────────────────────────────
+
+// stepXSSScanning collects parameterised URLs from the live URL file and
+// feeds them to Dalfox. This is the final step; it never signals cancellation.
+func stepXSSScanning(c *Ctx) {
+	if c.State.IsStepCompleted("xss_scanning") {
+		logger.Section("Step 20: XSS Scanning (Dalfox) [RESUMED — skipping]")
+		return
+	}
+
+	if !c.SkipDalfox {
+		logger.Section("Step 20: XSS Scanning (Dalfox)")
+		logger.SubStep("Collecting parameterized URLs from live URLs...")
+
+		paramCount := collectParamURLsFromFile(c.F.AllURLsLive, c.F.ParamURLsFile)
+
+		if paramCount > 0 {
+			logger.Info("  Found %d parameterized URLs to test", paramCount)
+			logger.SubStep("Running Dalfox on parameterized URLs...")
+
+			if err := runWithSkip(c, "dalfox", func(sCtx context.Context) error {
+				return c.Tb.RunDalfox(sCtx, c.F.ParamURLsFile, c.F.DalfoxOut)
+			}); err != nil {
+				if err == ErrToolSkipped {
+					logger.Info("  Dalfox skipped")
+				} else {
+					c.StateMgr.MarkStepFailed(c.State, "xss_scanning", err)
+					logger.Warning("Dalfox failed: %v", err)
+				}
+			} else {
+				if c.ScanID > 0 {
+					count, _ := utils.ParseDalfoxOutput(c.ScanID, c.F.DalfoxOut)
+					if count > 0 {
+						logger.Success("  Found %d XSS vulnerabilities!", count)
+					} else {
+						logger.Info("  No XSS vulnerabilities found")
+					}
+				}
+			}
+		} else {
+			logger.Info("  No parameterized URLs found to test for XSS")
+		}
+		c.StateMgr.MarkStepComplete(c.State, "xss_scanning")
+	} else {
+		logger.Section("Step 20: Skipping Dalfox (--skip-dalfox)")
+		c.StateMgr.MarkStepComplete(c.State, "xss_scanning")
+	}
+}
+
+// ─────────────────────────────────────────────────────────────
 // Internal — send notifications for critical/high findings
 // ─────────────────────────────────────────────────────────────
 
@@ -145,6 +238,22 @@ func sendVulnNotifications(c *Ctx, base notify.Finding) {
 				URL:         v.URL,
 				TemplateID:  v.TemplateID,
 				Timestamp:   base.Timestamp,
+			})
+		}
+	}
+}
+
+func sendTakeoverNotifications(c *Ctx) {
+	vulns, _ := database.GetVulnerabilities(c.ScanID)
+	for _, v := range vulns {
+		if v.TemplateID == "subdomain-takeover" {
+			c.Notifier.SendFinding(notify.Finding{
+				Target:    c.Domain,
+				Type:      "subdomain-takeover",
+				Name:      v.Name,
+				Severity:  "critical",
+				URL:       v.Host,
+				Timestamp: time.Now(),
 			})
 		}
 	}
